BUGFIX: Use size_t for array sizes instead of int
Using plain integers for the number of elements in arrays quickly
leads to overflows when using larger trace files. The result is an
error message when loading large traces although the trace file is
correct and enough memory is available. Using size_t for array sizes
fixes the problem.

+++ src/buffer.c
@@ -17,7 +17,7 @@
 
 #include "buffer.h"
 
-int check_buffer_grow(void** buffer, int elem_size, int used, int* free, int prealloc_elems)
+int check_buffer_grow(void** buffer, size_t elem_size, size_t used, size_t* free, size_t prealloc_elems)
 {
 	void* ptr;
 
@@ -33,7 +33,7 @@ int check_buffer_grow(void** buffer, int elem_size, int used, int* free, int pre
 	return 0;
 }
 
-int add_buffer_grow(void** buffer, void* elem, int elem_size, int* used, int* free, int prealloc_elems)
+int add_buffer_grow(void** buffer, void* elem, size_t elem_size, size_t* used, size_t* free, size_t prealloc_elems)
 {
 	if(check_buffer_grow(buffer, elem_size, *used, free, prealloc_elems))
 		return 1;
