+++ src/buffer.h
@@ -21,7 +21,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-int check_buffer_grow(void** buffer, int elem_size, int used, int* free, int prealloc_elems);
-int add_buffer_grow(void** buffer, void* elem, int elem_size, int* used, int* free, int prealloc_elems);
+int check_buffer_grow(void** buffer, size_t elem_size, size_t used, size_t* free, size_t prealloc_elems);
+int add_buffer_grow(void** buffer, void* elem, size_t elem_size, size_t* used, size_t* free, size_t prealloc_elems);
 
 #endif
